requirements.txt
streamlit
apscheduler
groq
requests
python-dotenv
duckduckgo-search
plotly
pyvis
fpdf2
wordcloud
matplotlib
networkx
